day2_a1/backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
langchain
langchain-community
langchain-groq
langchain-huggingface
chromadb
sentence-transformers
python-dotenv
jinja2
python-docx
langchain-experimental
pandas
